+++ Game/src/sys/System.h
@@ -6,7 +6,7 @@ class System
 public:
 	explicit System(entt::registry& registry) : mRegistry(registry) {}
 
-	virtual void update(float dt) {};
+	virtual void update(double dt) {};
 
 	static void setPause(bool pause) { sPause = pause; }
